# openapi/public-openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityEntry": {
        "description": "Recent activity entry",
        "properties": {
          "action": {
            "description": "Action type (stored, evict)",
            "title": "Action",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Conversation ID",
            "title": "Conversation Id"
          },
          "id": {
            "description": "Activity ID",
            "title": "Id",
            "type": "string"
          },
          "preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Message preview",
            "title": "Preview"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Message role (user, assistant)",
            "title": "Role"
          },
          "timestamp": {
            "description": "Activity timestamp",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "timestamp",
          "action"
        ],
        "title": "ActivityEntry",
        "type": "object"
      },
...
      "CollectionInfo": {
        "description": "Collection metadata.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "dimension": {
            "title": "Dimension",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "count",
          "dimension"
        ],
        "title": "CollectionInfo",
        "type": "object"
      },
...
      "SemanticMemoryStatsResponse": {
        "description": "Semantic memory statistics.",
        "properties": {
          "avg_retrieval_latency_ms": {
            "title": "Avg Retrieval Latency Ms",
            "type": "number"
          },
          "collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionInfo"
            },
            "title": "Collections",
            "type": "array"
          },
          "index_size_mb": {
            "title": "Index Size Mb",
            "type": "number"
          },
          "retrieval_quality_percent": {
            "default": 0.0,
            "title": "Retrieval Quality Percent",
            "type": "number"
          },
          "total_vectors": {
            "title": "Total Vectors",
            "type": "integer"
          }
        },
        "required": [
          "total_vectors",
          "collections",
          "index_size_mb",
          "avg_retrieval_latency_ms"
        ],
        "title": "SemanticMemoryStatsResponse",
        "type": "object"
      },
...
      "WorkingMemoryStatsResponse": {
        "description": "Working memory statistics.",
        "properties": {
          "active_items": {
            "title": "Active Items",
            "type": "integer"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "eviction_rate_per_min": {
            "title": "Eviction Rate Per Min",
            "type": "number"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/ActivityEntry"
            },
            "title": "Recent Activity",
            "type": "array"
          },
          "ttl_utilization_percent": {
            "title": "Ttl Utilization Percent",
            "type": "number"
          },
          "utilization_percent": {
            "title": "Utilization Percent",
            "type": "number"
          }
        },
        "required": [
          "active_items",
          "capacity",
          "utilization_percent",
          "ttl_utilization_percent",
          "eviction_rate_per_min",
          "recent_activity"
        ],
        "title": "WorkingMemoryStatsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AICO Backend REST API with clean architecture",
    "title": "AICO Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/memory/semantic/stats": {
      "get": {
        "description": "Get semantic memory (ChromaDB) statistics.\n\nReturns:\n    - total_vectors: Total vector count across collections\n    - collections: List of collections with counts and dimensions\n    - index_size_mb: Total index size in megabytes\n    - avg_retrieval_latency_ms: Average query latency",
        "operationId": "get_semantic_stats_api_v1_memory_semantic_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticMemoryStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Semantic Stats",
        "tags": [
          "memory",
          "memory"
        ]
      }
    },
    "/api/v1/memory/working/stats": {
      "get": {
        "description": "Get working memory (LMDB) statistics.\n\nReturns:\n    - active_items: Current number of items in memory\n    - capacity: Maximum capacity\n    - utilization_percent: Memory utilization percentage\n    - ttl_utilization_percent: TTL-based utilization\n    - eviction_rate_per_min: Items evicted per minute\n    - recent_activity: Recent read/write/evict operations",
        "operationId": "get_working_stats_api_v1_memory_working_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingMemoryStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Working Stats",
        "tags": [
          "memory",
          "memory"
        ]
      }
    },
...
  }
}